+++ web_poet/fields.py
@@ -5,7 +5,7 @@ into separate Page Object methods / properties.
 import inspect
 from contextlib import suppress
 from functools import update_wrapper, wraps
-from typing import Callable, Dict, List, Optional, Type, TypeVar
+from typing import Callable, Dict, List, Optional, Tuple, Type, TypeVar
 
 import attrs
 from itemadapter import ItemAdapter
@@ -40,9 +40,12 @@ class FieldsMixin:
         # the base class. This is done by making decorator write to a
         # temporary location, and then merging it all on subclass creation.
         this_class_fields = getattr(cls, _FIELDS_INFO_ATTRIBUTE_WRITE, {})
-        base_class_fields = getattr(cls, _FIELDS_INFO_ATTRIBUTE_READ, {})
-        if base_class_fields or this_class_fields:
-            fields = {**base_class_fields, **this_class_fields}
+        base_fields = {}
+        for base_class in cls.__bases__:
+            fields = getattr(base_class, _FIELDS_INFO_ATTRIBUTE_READ, {})
+            base_fields.update(fields)
+        if base_fields or this_class_fields:
+            fields = {**base_fields, **this_class_fields}
             setattr(cls, _FIELDS_INFO_ATTRIBUTE_READ, fields)
             with suppress(AttributeError):
                 delattr(cls, _FIELDS_INFO_ATTRIBUTE_WRITE)
@@ -77,11 +80,9 @@ def field(
                 raise TypeError(
                     f"@field decorator must be used on methods, {method!r} is decorated instead"
                 )
-            method = self._processed(method)
-            if cached:
-                self.unbound_method = cached_method(method)
-            else:
-                self.unbound_method = method
+            self.original_method = method
+            self.unbound_method = None
+            self.processors: List[Tuple[Callable, bool]] = []
 
         def __set_name__(self, owner, name):
             if not hasattr(owner, _FIELDS_INFO_ATTRIBUTE_WRITE):
@@ -91,16 +92,27 @@ def field(
             getattr(owner, _FIELDS_INFO_ATTRIBUTE_WRITE)[name] = field_info
 
         def __get__(self, instance, owner=None):
+            if self.unbound_method is None:
+                for processor in out or []:
+                    sig = inspect.signature(processor)
+                    self.processors.append((processor, "instance" in sig.parameters))
+                method = self._processed(self.original_method, instance)
+                if cached:
+                    self.unbound_method = cached_method(method)
+                else:
+                    self.unbound_method = method
+
             return self.unbound_method(instance)
 
-        @staticmethod
-        def _process(value):
-            if out:
-                for processor in out:
+        def _process(self, value, instance):
+            for processor, takes_instance in self.processors:
+                if takes_instance:
+                    value = processor(value, instance=instance)
+                else:
                     value = processor(value)
             return value
 
-        def _processed(self, method):
+        def _processed(self, method, instance):
             """Returns a wrapper for method that calls processors on its result"""
             if inspect.iscoroutinefunction(method):
 
@@ -108,7 +120,7 @@ def field(
                     validation_item = args[0]._validate_input()
                     if validation_item is not None:
                         return getattr(validation_item, method.__name__)
-                    return self._process(await method(*args, **kwargs))
+                    return self._process(await method(*args, **kwargs), instance)
 
             else:
 
@@ -116,7 +128,7 @@ def field(
                     validation_item = args[0]._validate_input()
                     if validation_item is not None:
                         return getattr(validation_item, method.__name__)
-                    return self._process(method(*args, **kwargs))
+                    return self._process(method(*args, **kwargs), instance)
 
             return wraps(method)(processed)